Cache: Direct Mapping Cache
A Direct Map Cache implementation with
1-word blocks. Byte Addresses converted to
Block Addresses and matched against TIO
and valid bit in the cache.

If the cache does not contain the data at
the Block Address, the data is fetched from
Disk and stored in the cache, replacing the
data previously at that Block Address.

No cache optimizations included.