Graph: Shortest Path Implementation
This file contains an implementation for finding the shortest path in a graph from one node to another. The shortest distances are tracked in the paths array, and the optimal edge from one vertex to another is tracked in the edgeTo array. The adjacency matrix is maintained in a HashMap.