Graph: Eulerian Paths and Circuits
This file contains an implementation of the Eulerian Paths and Circuit detection algorithm. An Eulerian Path is a path in the graph that uses every edge exactly once. An Eulerian Circuit is a path in the graph that uses every edge exactly once and returns to the starting node at the end of the path.