Graph: Prim's Algorithm
This file contains an implementation of Prim's Algorithm using an Indexed Priority Queue to sort edges by their weight and update better paths to a vertex, if one exists. Prim's Algorithm finds the Minimum Spanning Tree of a graph. An MST is a tree that contains V - 1 edges, connects all vertices, and costs the least in terms of the sum of edge weight values. MSTs are not guaranteed to be unique.